+++ modules/gui/desktop/hyprland/apps/panel.nix
@@ -21,7 +21,7 @@ in
     programs.waybar = {
       enable = true;
       systemd.enable = true;
-      package = pkgs.waypkgs.waybar;
+      package = pkgs.unstable.waybar;
       style = files.hyprland.waybar;
 
       # Panel